+++ app/controllers/komet_dashboard_controller.rb
@@ -30,7 +30,7 @@ class KometDashboardController < ApplicationController
     include TaxonomyConcern, ConceptConcern, InstrumentationConcern, SearchApis
     include CommonController, TaxonomyHelper
 
-    TAXONOMY_CHILD_PAGE_SIZE = 1000
+    TAXONOMY_CHILD_PAGE_SIZE = 250
 
     #before_action :setup_routes, :setup_constants, :only => [:dashboard] #todo Ask Tim why is this in app controller too?  OK to whack?
     after_filter :byte_size unless Rails.env.production?
@@ -45,22 +45,8 @@ class KometDashboardController < ApplicationController
     # If the tree is reversed so we are searching for parents of this node is identified in the request params with the key :parent_search (true/false)
     # If the parent of this node was already doing a reverse search is identified in the request params with the key :parent_reversed (true/false)
     # [Object] view_params - various parameters related to the view filters the user wants to apply - see full definition comment at top of komet_dashboard_controller file
-    #@return [json] the tree nodes to insert into the tree at the parent node passed in the request
+    # @return [json] the tree nodes to insert into the tree at the parent node passed in the request
     def load_tree_data
-
-        # roles = session[Roles::SESSION_ROLES_ROOT][Roles::SESSION_USER_ROLES]
-        # if(roles.include?(Roles::DEV_SUPER_USER))
-        #   #do something
-        # end
-        # CRIS, test flash code
-        # resp = CoordinateRest.get_coordinate(action: CoordinateRestActions::ACTION_COORDINATES_TOKEN, additional_req_params: {foo: 'faa'})
-        # resp.rest_exception.flash_error
-        # resp.flash_error   (both work)
-        # if resp.is_a? CommonRest::UnexpectedResponse
-        #   resp.flash_error
-        # end
-        #
-        # resp.flash_error if resp.respond_to? :flash_error
 
         # TODO - see if some of these can just be referenced everywhere as params[*] instead of passing variables to all sub functions
         selected_concept_id = params[:concept_id]
@@ -68,6 +54,7 @@ class KometDashboardController < ApplicationController
         parent_reversed = params[:parent_reversed]
         tree_walk_levels = params[:tree_walk_levels]
         multi_path = params[:multi_path]
+        next_page = params[:next_page]
         @view_params = check_view_params(params[:view_params])
 
         # check to make the number of levels to walk the tree was passed in
@@ -82,18 +69,23 @@ class KometDashboardController < ApplicationController
             multi_path = true
         end
 
-        tree_nodes = populate_tree(selected_concept_id, parent_search, parent_reversed, tree_walk_levels, multi_path)
+        tree_nodes = populate_tree(selected_concept_id, parent_search, parent_reversed, tree_walk_levels, multi_path, next_page)
 
         render json: tree_nodes
     end
 
-    def populate_tree(selected_concept_id, parent_search, parent_reversed, tree_walk_levels, multi_path)
+    def populate_tree(selected_concept_id, parent_search, parent_reversed, tree_walk_levels, multi_path, next_page = nil)
 
         coordinates_token = session[:coordinates_token].token
         root = selected_concept_id.eql?('#')
 
         additional_req_params = {coordToken: coordinates_token, sememeMembership: true, maxPageSize: TAXONOMY_CHILD_PAGE_SIZE}
         additional_req_params.merge!(@view_params)
+
+        # check to see if we are getting the next page of child results, if so add the pageNum parameter
+        if next_page != nil
+            additional_req_params[:pageNum] = next_page
+        end
 
         if boolean(parent_search)
             tree_walk_levels = 100
@@ -307,7 +299,16 @@ class KometDashboardController < ApplicationController
 
             if raw_node[:next_page]
 
-                tree_nodes << {id: get_next_id, concept_id: raw_node[:id], text: raw_node[:text], next_page: raw_node[:next_page], children: false, parent_reversed: false, parent_search: false, icon: 'komet-tree-node-icon komet-tree-node-defined', a_attr: {class: 'komet-tree-warning', 'aria-label' => raw_node[:text]}}
+                tree_nodes << {id: get_next_id,
+                               concept_id: raw_node[:id],
+                               text: raw_node[:text],
+                               next_page: raw_node[:next_page],
+                               children: false,
+                               parent_reversed: false,
+                               parent_search: false,
+                               icon: 'komet-tree-node-icon fa fa-info-circle',
+                               a_attr: {class: 'komet-tree-warning', 'aria-label' => raw_node[:text]}
+                }
                 next
             end