+++ ChangeLog.md
@@ -6,6 +6,7 @@ where provided, and the git commit history.
 * 2017/07/?? - 4.16 - PENDING
     * Fixed javascript error when saving mapsets (Jazz: 550838)
     * Updated mapsets to use new item paging API changes
+    * Taxonomy tree will now load additional pages of children, 250 at a time. (Jazz: 554508)
 
 * 2017/07/13 - 4.15
     * Fixed Add Property so the dialog will close properly if you click Add, get an error, and then click Cancel.